requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
attrs==25.3.0
frozenlist==1.7.0
idna==3.10
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.6.3
prompt-toolkit==3.0.51
propcache==0.3.2
pyaes==1.6.1
pyasn1==0.6.1
pygments==2.19.2
pyyaml==6.0.2
rich==14.1.0
rsa==4.9.1
tabulate==0.9.0
telethon==1.40.0
tqdm==4.67.1
typing-extensions==4.14.1
wcwidth==0.2.13
yarl==1.20.1
